+++ src/Game/GameObjects/Eagle.h
@@ -24,7 +24,7 @@ namespace BatleCity
 		Eagle(const glm::vec2& positiion, const glm::vec2& size, const float rotation, const float layer);
 
 		void render() const override;
-		void update(const uint64_t delta) override;
+		void update(const double delta) override;
 
 	private:
